Updated Settings, URLs, and generate_key
In the settings file, I added the rest_framework to the installed apps, as well as set up the backend for the serializer.
URLs, I added our first url, and the url to our API
Generate_key, this file generates a  secret key made of random keys. I replaced the 'secret key' in the settings with a reference to the code in this file.