Created unit model and REST API
Added files; urls, serializers, json_views, api_urls, and initial_data
The serializers file, makes the data in the model able to be filtered.
The model i created is for storing GPS units.
This is part of the REST framework that will translate/filter the json data passed through
api_urls, this file contains the url to link the json_view file and filter
the initial_data file contains json data from the database. This is part of the datadump function in PyCharm