Added URLs, templates, and implemented bootstrap
All of the changes made so far in this project have been experimental at best.
I created a layout and some urls and linked them to the bootstrap CSS/JS files.
I also added bootstrap to the installed apps in the settings.py file. The base file
is for blocks of code to be used in other templates.